Pandar OT 128E4X Initial support (#25)
* wip. pandar ot


* wip. missing correction


* pandar128e4x. add calibration file


* pandar_ot. initial working version


* pandar_ot. clean up


* nebula. apply clang format


---------

+++ nebula_drivers/include/hesai/decoders/hesai_scan_decoder.hpp
@@ -24,7 +24,7 @@ protected:
   std::shared_ptr<drivers::HesaiSensorConfiguration> sensor_configuration_;
   std::shared_ptr<drivers::HesaiCalibrationConfiguration> sensor_calibration_;
 
-  static inline double deg2rad(double degrees) { return degrees * M_PI / 180.0; }
+  static inline float deg2rad(double degrees) { return degrees * M_PI / 180.0; }
 
 public:
   HesaiScanDecoder(HesaiScanDecoder && c) = delete;
